[5285] Extract the creation of RepresentationEventProcessors and the dispatch of IInput from the EditingContextEventProcessor
Bug: https://github.com/eclipse-sirius/sirius-web/issues/5285
