+++ packages/core/backend/sirius-components-collaborative/src/main/java/org/eclipse/sirius/components/collaborative/editingcontext/EditingContextEventProcessor.java
@@ -12,34 +12,29 @@
  *******************************************************************************/
 package org.eclipse.sirius.components.collaborative.editingcontext;
 
-import java.time.Duration;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
 import java.util.logging.Level;
 
 import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.Timer;
 import org.eclipse.sirius.components.collaborative.api.ChangeDescription;
-import org.eclipse.sirius.components.collaborative.api.IEditingContextEventHandler;
 import org.eclipse.sirius.components.collaborative.api.IEditingContextEventProcessor;
-import org.eclipse.sirius.components.collaborative.api.IInputPostProcessor;
-import org.eclipse.sirius.components.collaborative.api.IInputPreProcessor;
+import org.eclipse.sirius.components.collaborative.editingcontext.api.IInputDispatcher;
 import org.eclipse.sirius.components.collaborative.api.IRepresentationEventProcessor;
-import org.eclipse.sirius.components.collaborative.editingcontext.api.IChangeDescriptionListener;
-import org.eclipse.sirius.components.collaborative.api.IRepresentationEventProcessorComposedFactory;
+import org.eclipse.sirius.components.collaborative.editingcontext.api.IRepresentationEventProcessorProvider;
 import org.eclipse.sirius.components.collaborative.api.Monitoring;
-import org.eclipse.sirius.components.collaborative.dto.DeleteRepresentationInput;
+import org.eclipse.sirius.components.collaborative.editingcontext.api.IChangeDescriptionListener;
+import org.eclipse.sirius.components.collaborative.editingcontext.api.IEditingContextEventProcessorExecutorServiceProvider;
 import org.eclipse.sirius.components.collaborative.representations.api.IRepresentationEventProcessorRegistry;
-import org.eclipse.sirius.components.core.api.ErrorPayload;
 import org.eclipse.sirius.components.core.api.IEditingContext;
 import org.eclipse.sirius.components.core.api.IInput;
 import org.eclipse.sirius.components.core.api.IPayload;
-import org.eclipse.sirius.components.core.api.IRepresentationInput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import reactor.core.Disposable;
@@ -50,7 +45,6 @@ import reactor.core.publisher.Sinks;
 import reactor.core.publisher.Sinks.EmitResult;
 import reactor.core.publisher.Sinks.Many;
 import reactor.core.publisher.Sinks.One;
-import reactor.core.scheduler.Schedulers;
 
 /**
  * Handles all the inputs which concern a particular editing context one at a time, in order of arrival, and in a
@@ -65,19 +59,17 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
 
     public static final String INPUT = "INPUT";
 
-    private static final String LOG_TIMING_FORMAT = "%1$6s";
-
     private final Logger logger = LoggerFactory.getLogger(EditingContextEventProcessor.class);
 
     private final IEditingContext editingContext;
+
+    private final IInputDispatcher inputDispatcher;
 
     private final IRepresentationEventProcessorRegistry representationEventProcessorRegistry;
 
     private final IChangeDescriptionListener changeDescriptionListener;
 
-    private final List<IEditingContextEventHandler> editingContextEventHandlers;
-
-    private final IRepresentationEventProcessorComposedFactory representationEventProcessorComposedFactory;
+    private final IRepresentationEventProcessorProvider representationEventProcessorProvider;
 
     private final Many<IPayload> sink = Sinks.many().multicast().directBestEffort();
 
@@ -89,23 +81,18 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
 
     private final Disposable changeDescriptionDisposable;
 
-    private final List<IInputPreProcessor> inputPreProcessors;
-
-    private final List<IInputPostProcessor> inputPostProcessors;
-
     private final MeterRegistry meterRegistry;
 
-    public EditingContextEventProcessor(EditingContextEventProcessorParameters parameters) {
-        this.editingContext = parameters.editingContext();
-        this.representationEventProcessorRegistry = parameters.representationEventProcessorRegistry();
-        this.editingContextEventHandlers = parameters.editingContextEventHandlers();
-        this.representationEventProcessorComposedFactory = parameters.representationEventProcessorComposedFactory();
-        this.changeDescriptionListener = parameters.changeDescriptionListener();
-        this.executorService = parameters.executorServiceProvider().getExecutorService(this.editingContext);
-        this.inputPreProcessors = parameters.inputPreProcessors();
-        this.inputPostProcessors = parameters.inputPostProcessors();
+    public EditingContextEventProcessor(IEditingContextEventProcessorExecutorServiceProvider executorServiceProvider, IEditingContext editingContext, IRepresentationEventProcessorRegistry representationEventProcessorRegistry, IChangeDescriptionListener changeDescriptionListener, IInputDispatcher inputDispatcher, IRepresentationEventProcessorProvider representationEventProcessorProvider, MeterRegistry meterRegistry) {
+        this.editingContext = Objects.requireNonNull(editingContext);
+        this.representationEventProcessorRegistry = Objects.requireNonNull(representationEventProcessorRegistry);
+        this.changeDescriptionListener = Objects.requireNonNull(changeDescriptionListener);
+        this.executorService = executorServiceProvider.getExecutorService(this.editingContext);
+        this.inputDispatcher = Objects.requireNonNull(inputDispatcher);
+        this.representationEventProcessorProvider = Objects.requireNonNull(representationEventProcessorProvider);
+        this.meterRegistry = Objects.requireNonNull(meterRegistry);
         this.changeDescriptionDisposable = this.setupChangeDescriptionSinkConsumer();
-        this.meterRegistry = parameters.meterRegistry();
+
     }
 
     private Disposable setupChangeDescriptionSinkConsumer() {
@@ -125,15 +112,14 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
         Timer.Sample handleTimer = Timer.start(this.meterRegistry);
         if (this.executorService.isShutdown()) {
             this.logger.warn("Handler for editing context {} is shutdown", this.editingContext.getId());
-            handleTimer.stop(this.meterRegistry.timer(Monitoring.EVENT_HANDLER, INPUT, input.getClass().getSimpleName(),
-                    "inputId", input.id().toString()));
+            handleTimer.stop(this.meterRegistry.timer(Monitoring.EVENT_HANDLER, INPUT, input.getClass().getSimpleName(), "inputId", input.id().toString()));
             return Mono.empty();
         }
 
         this.logger.trace(input.toString());
 
         One<IPayload> payloadSink = Sinks.one();
-        Future<?> future = this.executorService.submit(() -> this.doHandle(payloadSink, input));
+        Future<?> future = this.executorService.submit(() -> this.inputDispatcher.dispatch(this.executorService, payloadSink, this.canBeDisposedSink, this.changeDescriptionSink, this.editingContext, input));
         try {
             // Block until the event has been processed
             future.get();
@@ -148,123 +134,15 @@ public class EditingContextEventProcessor implements IEditingContextEventProcess
                 .doOnError(throwable -> this.logger.warn(throwable.getMessage(), throwable));
     }
 
-    /**
-     * Finds the proper event handler to perform the task matching the given input event.
-     *
-     * @param payloadSink
-     *         The sink to publish payload
-     * @param input
-     *         The input event
-     */
-    private void doHandle(One<IPayload> payloadSink, IInput input) {
-        this.logger.trace("Input received: {}", input);
-        long start = System.currentTimeMillis();
-
-        AtomicReference<IInput> inputAfterPreProcessing = new AtomicReference<>(input);
-        this.inputPreProcessors.forEach(preProcessor -> inputAfterPreProcessing.set(preProcessor.preProcess(this.editingContext, inputAfterPreProcessing.get(), this.changeDescriptionSink)));
-
-        if (inputAfterPreProcessing.get() instanceof IRepresentationInput representationInput) {
-            this.handleRepresentationInput(payloadSink, representationInput);
-        } else {
-            this.handleInput(payloadSink, inputAfterPreProcessing.get());
-        }
-
-        this.inputPostProcessors.forEach(postProcessor -> postProcessor.postProcess(this.editingContext, inputAfterPreProcessing.get(), this.changeDescriptionSink));
-
-        long end = System.currentTimeMillis();
-        this.logger.atDebug()
-                .setMessage("EditingContext {}: {}ms to handle the {} with id {}")
-                .addArgument(this.editingContext.getId())
-                .addArgument(() -> String.format(LOG_TIMING_FORMAT, end - start))
-                .addArgument(input.getClass().getSimpleName())
-                .addArgument(input.id())
-                .log();
-    }
-
-    private void handleInput(One<IPayload> payloadSink, IInput input) {
-        if (input instanceof DeleteRepresentationInput deleteRepresentationInput) {
-            this.disposeRepresentation(deleteRepresentationInput.representationId());
-        }
-
-        Optional<IEditingContextEventHandler> optionalEditingContextEventHandler = this.editingContextEventHandlers.stream()
-                .filter(handler -> handler.canHandle(this.editingContext, input))
-                .findFirst();
-
-        if (optionalEditingContextEventHandler.isPresent()) {
-            IEditingContextEventHandler editingContextEventHandler = optionalEditingContextEventHandler.get();
-            editingContextEventHandler.handle(payloadSink, this.changeDescriptionSink, this.editingContext, input);
-        } else {
-            this.logger.warn("No handler found for event: {}", input);
-            payloadSink.tryEmitValue(new ErrorPayload(input.id(), "No handler found for event: " + input.toString()));
-        }
-    }
-
-    private void handleRepresentationInput(One<IPayload> payloadSink, IRepresentationInput representationInput) {
-        Optional<IRepresentationEventProcessor> optionalRepresentationEventProcessor = this.acquireRepresentationEventProcessor(representationInput.representationId(), representationInput);
-
-        if (optionalRepresentationEventProcessor.isPresent()) {
-            IRepresentationEventProcessor representationEventProcessor = optionalRepresentationEventProcessor.get();
-            representationEventProcessor.handle(payloadSink, this.changeDescriptionSink, representationInput);
-        } else {
-            this.logger.warn("No representation event processor found for event: {}", representationInput);
-        }
-    }
 
     @Override
     public Optional<IRepresentationEventProcessor> acquireRepresentationEventProcessor(String representationId, IInput input) {
-        var getRepresentationEventProcessorSample = Timer.start(this.meterRegistry);
-
-        var optionalRepresentationEventProcessor = Optional.ofNullable(this.representationEventProcessorRegistry.get(this.editingContext.getId(), representationId))
-                .map(RepresentationEventProcessorEntry::getRepresentationEventProcessor);
-
-        if (optionalRepresentationEventProcessor.isEmpty()) {
-            optionalRepresentationEventProcessor = this.representationEventProcessorComposedFactory.createRepresentationEventProcessor(this.editingContext, representationId);
-            if (optionalRepresentationEventProcessor.isPresent()) {
-                var representationEventProcessor = optionalRepresentationEventProcessor.get();
-
-                Disposable subscription = representationEventProcessor.canBeDisposed()
-                        .delayElements(Duration.ofSeconds(5))
-                        .publishOn(Schedulers.fromExecutorService(this.executorService))
-                        .subscribe(canBeDisposed -> {
-                            if (canBeDisposed.booleanValue() && representationEventProcessor.getSubscriptionManager().isEmpty()) {
-                                this.disposeRepresentation(representationId);
-                            } else {
-                                this.logger.trace("Stopping the disposal of the representation event processor {}", representationId);
-                            }
-                        }, throwable -> this.logger.warn(throwable.getMessage(), throwable));
-
-                var representationEventProcessorEntry = new RepresentationEventProcessorEntry(representationEventProcessor, subscription);
-                this.representationEventProcessorRegistry.put(this.editingContext.getId(), representationId, representationEventProcessorEntry);
-            } else {
-                this.logger.debug("The representation with the id {} does not exist", representationId);
-            }
-        } else {
-            var timer = this.meterRegistry.timer(Monitoring.TIMER_CREATE_REPRESENATION_EVENT_PROCESSOR,
-                    "editingContext", this.editingContext.getId(),
-                    "input", input.getClass().getSimpleName(),
-                    "representationId", representationId);
-            getRepresentationEventProcessorSample.stop(timer);
-        }
-
-        this.logger.trace("Representation event processors count: {}", this.representationEventProcessorRegistry.values(this.editingContext.getId()).size());
-        return optionalRepresentationEventProcessor;
+        return this.representationEventProcessorProvider.acquireRepresentationEventProcessor(this.executorService, this.canBeDisposedSink, this.editingContext, representationId, input);
     }
 
     @Override
     public List<IRepresentationEventProcessor> getRepresentationEventProcessors() {
         return this.representationEventProcessorRegistry.values(this.editingContext.getId());
-    }
-
-    private void disposeRepresentation(String representationId) {
-        this.representationEventProcessorRegistry.disposeRepresentation(this.editingContext.getId(), representationId);
-
-        if (this.representationEventProcessorRegistry.values(this.editingContext.getId()).isEmpty()) {
-            EmitResult emitResult = this.canBeDisposedSink.tryEmitNext(Boolean.TRUE);
-            if (emitResult.isFailure()) {
-                String pattern = "An error has occurred while emitting that the processor can be disposed: {}";
-                this.logger.warn(pattern, emitResult);
-            }
-        }
     }
 
     @Override
